third_party/blink/web_tests/virtual/interest-target-no-partial-interest/external/wpt/html/semantics/the-button-element/interest-target/interesttarget-popover-focus.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] It should be possible to move focus into target popover
  assert_equals: expected Element node <button id="contained">button</button> but got Element node <button id="el" interesttarget="target">Button</button>
Harness: the test ran to completion.
